# Add check-latest flag