--- a/src/distributions/base-installer.ts
+++ b/src/distributions/base-installer.ts
@@ -12,6 +12,7 @@ export abstract class JavaBase {
   protected architecture: string;
   protected packageType: string;
   protected stable: boolean;
+  protected checkLatest: boolean;
 
   constructor(protected distribution: string, installerOptions: JavaInstallerOptions) {
     this.http = new httpm.HttpClient('actions/setup-java', undefined, {
@@ -24,6 +25,7 @@ export abstract class JavaBase {
     ));
     this.architecture = installerOptions.architecture;
     this.packageType = installerOptions.packageType;
+    this.checkLatest = !!installerOptions.checkLatest;
   }
 
   protected abstract downloadTool(javaRelease: JavaDownloadRelease): Promise<JavaInstallerResults>;
@@ -31,13 +33,17 @@ export abstract class JavaBase {
 
   public async setupJava(): Promise<JavaInstallerResults> {
     let foundJava = this.findInToolcache();
-    if (foundJava) {
+    if (foundJava && !this.checkLatest) {
       core.info(`Resolved Java ${foundJava.version} from tool-cache`);
     } else {
       core.info(`Java ${this.version} was not found in tool-cache. Trying to download...`);
       const javaRelease = await this.findPackageForDownload(this.version);
-      foundJava = await this.downloadTool(javaRelease);
-      core.info(`Java ${foundJava.version} was downloaded`);
+      if (foundJava?.version != javaRelease.version) {
+        foundJava = await this.downloadTool(javaRelease);
+        core.info(`Java ${foundJava.version} was downloaded`);
+      }
+
+      core.info(`Java ${foundJava.version} was resolved`);
     }
 
     core.info(`Setting Java ${foundJava.version} as the default`);