--- a/src/distributions/base-models.ts
+++ b/src/distributions/base-models.ts
@@ -2,6 +2,7 @@ export interface JavaInstallerOptions {
   version: string;
   architecture: string;
   packageType: string;
+  checkLatest?: boolean;
 }
 
 export interface JavaInstallerResults {