--- a/src/distributions/base-models.ts
+++ b/src/distributions/base-models.ts
@@ -2,7 +2,7 @@ export interface JavaInstallerOptions {
   version: string;
   architecture: string;
   packageType: string;
-  checkLatest?: boolean;
+  checkLatest: boolean;
 }
 
 export interface JavaInstallerResults {