--- a/src/setup-java.ts
+++ b/src/setup-java.ts
@@ -1,6 +1,6 @@
 import * as core from '@actions/core';
 import * as auth from './auth';
-
+import { getBooleanInput } from './util';
 import * as constants from './constants';
 import * as path from 'path';
 import { getJavaDistribution } from './distributions/distribution-factory';
@@ -13,13 +13,13 @@ async function run() {
     const architecture = core.getInput(constants.INPUT_ARCHITECTURE);
     const packageType = core.getInput(constants.INPUT_JAVA_PACKAGE);
     const jdkFile = core.getInput(constants.INPUT_JDK_FILE);
-    const checkLatest = core.getInput(constants.INPUT_CHECK_LATEST);
+    const checkLatest = getBooleanInput(constants.INPUT_CHECK_LATEST, false);
 
     const installerOptions: JavaInstallerOptions = {
       architecture,
       packageType,
       version,
-      checkLatest: checkLatest ? false : checkLatest.toLowerCase() === 'true'
+      checkLatest
     };
 
     const distribution = getJavaDistribution(distributionName, installerOptions, jdkFile);