**Description:**
In scope of this pull request we add support for check-latest flag. 
**Related issue:** https://github.com/actions/virtual-environments-internal/issues/1942

**Check list:**
- [ ] Mark if documentation changes are required.
- [x] Mark if tests were added or updated to cover the changes.