@giltene 2021-03-19
The default should be true, not false. If someone wants to avoid checking for latest update, that may be ok. But defaulting to that behavior is problematic for many common use cases of setup-java.