@dmitry-shibanov 2021-03-22
Hello. Thank you for your participation and help with improving setup-java. We prefer to keep check-latest flag as false for default cases, because we want to reduce task duration that is why we prefer to use our  image cache as much as possible. Related [comment](https://github.com/actions/setup-java/pull/132#issuecomment-801067914).