@giltene 2021-03-22
Is there some other place or forum where these discussions and decisions are being made? Who is the team making the decisions? Can I join that team? 
Given the dramatic impact proposed with this change, and in a few related changes that seem to be taking shape, a more transparent process that would allow significantly more user input prior to making final choices would be beneficial.