@konradpabjan 2021-03-22
@giltene 

> The default should be true, not false. If someone wants to avoid checking for latest update, that may be ok (although I don't see why anyone would choose to do that, ever)

It's important to keep in mind the use-case we're trying to optimize here for. Since late last year our hosted-images come with a few versions of Java from `adopt`. There are no versions of Java from `Zulu` that are cached for a while now (there used to be versions of `Zulu` that were cached but we switched over): https://github.com/actions/virtual-environments/blob/main/images/linux/Ubuntu2004-README.md#java

How often are these versions of `adopt` that are cached updated? Since October of last year, they've been updated 3 times and on average it takes ~4 days for our images to get updated with the latest versions and roll-out to everyone. Sometimes this update process takes more, sometimes it takes less. 

If a cached version of adopt is used vs downloaded from scratch, the time benefit is ~0s vs ~20s. If `check-latest` is set to true during these update times then the `setup-java` step would all of a sudden take a longer period of time (not ideal experience for users). Also the time benefit of this cannot be understated from a compute perspective for hosted runners that Github manages. 20 seconds doesn't sound like much, but if you factor in the millions of runs, and VMs in hosted that we have to spin up, re-image etc. 20 seconds each on a common setup action is a pretty big deal. By setting `check-latest` to `true` there would be no benefit for us.

So for hosted runners that use `setup-java`, the benefit for `check-latest` being `false` is pretty clear to me. If you're using another vendor then it makes no difference if `check-latest` is `false` or `true` on hosted since the distribution will have to be downloaded from scratch every single time since nothing else is cached.

For self-hosted runners there is a slightly different use case and I would argue that it makes sense to have `check-latest` set to `true` if you want the latest and greatest. However, we don't have a flag or env variable or anything that denotes something like `isSelfHosted`. The discussion has come up in a few other first party actions if we should have different behaviors for actions that are on hosted vs self-hosted and the answer that came up was no. This could be problematic for matrix builds for example that have some self-hosted runners and some hosted. So for consistency we just said no difference in behavior vs self-hosted and hosted across the board.

That being said, `check-latest` being `false` for self-hosted isn't all that bad and I think the pros still outweighs the cons (some might say this doesn't, and that's fair, but for us optimizing for hosted is orders of magnitude more important). Why?
- Currently `setup-java` V1 behaves as if `check-latest` is set to `false`. No breaking changes with `V2` here. This just gives users more flexibility
- I like to see it almost as an auto-update flag. Having your builds start failing all of a sudden because a java distribution auto-updated could prove problematic, so pinning to a cached version always is a more "stable" CI environment. If someone does want the latest version then that is something they should opt-in to.
- In `setup-node` there is also a `check-latest` option that has been around for a while, and it's worked great. It's set to `false` so for consistency and we with other first-party actions we should do the same: https://github.com/actions/setup-node/blob/main/action.yml#L12-L14 This allows us for 

> Is there some other place or forum where these discussions and decisions are being made?

The core specs for `setup-java` V2 are documented in this PR and for the most part that what we're following and implementing: https://github.com/actions/setup-java/pull/97
https://github.com/actions/setup-java/blob/main/docs/adrs/0000-v2-setup-java.md

This `check-latest` flag was something we figured, why not? Let's add this option in for v2 since it has valid use cases and there has been an ask for this feature for a while, we can squeeze it in and it will be a nice "cherry-on-top".

> Given the dramatic impact proposed with this change

For hosted this would have no change and it would clearly benefit us while for self-hosted it behaves just like V1 so IMO characterizing this change as dramatic is a bit of an overstatement