@ibauersachs 2021-03-23
> > Since when has (or had) the pre-installed Java version of the runner anything to do with this action? Does (or did) the action take advantage of the pre-installed Java version, i.e. something like if (version == 11 && distribution == 'adopt' && arch == 'x64') return;?
> 
> Currently it doesn't take advantage of any cached distributions because our images only cache some adopt distributions while V1 only is structured to pull Zulu distributions: https://github.com/actions/virtual-environments/blob/main/images/linux/Ubuntu2004-README.md#java

That just states that two random Java versions are installed in the image and that I could run `java` or `javac` without installing anything. If I wanted to do that, I wouldn't use `setup-java`.

> There was a time though when our hosted images cached Zulu distributions but we've switched over, see: [actions/virtual-environments#1057](https://github.com/actions/virtual-environments/issues/1057)
> Even when Zulu distributions were cached, our tool/directory structure wasn't set up to easily consume and switch between different cached versions so we didn't fully take advantage of this. Part of V2 involves actually using cached distributions on our images when possible to speed up setup and the work has already been done for that.

A cache is something you populate and reuse as long as it isn't outdated, based on a constraint (time, memory, ...). A pre-installed tool on a VM image is NOT a cache. At best you can pre-populate a cache in a VM image.

> > If the pre-installed version has nothing to do with what setup-java does (which I sincerely hope)
> 
> They do. I wholeheartedly disagree with notion that setup actions should avoid anything with pre-installed versions. There are overwhelming benefits. A good example is the work we did with setup-python v2 that we're now trying to bring over to `setup-java`. There are 6 versions of CPython that are cached on a hosted Image and 3 versions of PyPy (you can think of these as two different "distributions"): https://github.com/actions/virtual-environments/blob/main/images/linux/Ubuntu2004-README.md#python

Python is broken. It needs to be installed and cannot simply be downloaded into a directory and added to $PATH. Please don't use such a broken toolchain as an example.
And again: cache != preinstalled. But it seems that Actions abuses the word cache und thus causes my confusion.

> There are WAY more versions available than the 9 that are cached on each image and you can download those on the fly, the disadvantage though is that downloading and setting up a version dynamically can take a long time (over a minute on Windows for example!). The setup-python action has thus been optimized to use cached distributions when possible for the most possible versions. The cached versions have a slightly slower update cadence **but for the majority of users this is fine** and there are ways to get the latest and greatest if you're insistent on using the most up-to-date releases. You can read up how setup-java handles using pre-cached versions here: https://github.com/actions/setup-python#available-versions-of-python

So is python-versions an indication that there is soon going to be `GitHub OpenJDK for Actions`? If not, then the point is again that Python is broken.

> With setup-java you can imagine someone in V2 using `version == 11 && distribution == 'adopt` and our images have `11.0.10` ~~cached~~ **preinstalled**.

> Now let's assume `11.0.20` comes out. Now, realistically it will take some time to update and test our images with `11.0.20` and we update it eventually, but before that the actions shouldn't differ to download `11.0.20` and effectively disregard whatever is in the cache. **When possible we want to use it to make setup as fast as possible**.

Of course it should be fast, but then `11.0.20` needs to be available as soon as api.adoptopenjdk.net gives it back. Days later is too late. Not even `apt-get update && apt-get upgrade` would be so slow.

> From some of our other actions, users are very vocal about prioritizing faster runs as their biggest requests. If someone really wants `11.0.20` before it's on our image then they can specify `check-latest` as `true` with the option in this issue or just specify `version == 11.0.20`

Yes, hurray, I want faster runs too, especially on Macs! But a) not at the cost of having an arbitrary preinstalled version vs. what comes back from api.adoptopenjdk.net. and b) the time for a run is mostly spent in slow builds, probably due to random disk access, not in `setup-java`. In dnsjava/dnsjava, the JDK setup takes anywhere between 3 to 8s on Ubuntu, and it includes an (unnecessary) download. That is often faster than restoring the Maven cache. On Windows it's 7 to 12s and it's always faster than restoring the Maven cache.

I will definitely always set `check-latest: true`.

> > Instead, the actual core concern of a 20 second start can be addressed just as well (or to within fraction of a second of) by addressing #84 directly. I would be happy to do the work to address that one, and to commit to build it such that it will support caching for all distros to avoid distro favoritism arguments.
> 
> Solutions such as those outlined in #84 aren't optimal because `actions/cache` still involve a step that downloads from somewhere (although this in theory should be a little bit faster). The absolute fastest solution is to cache the most used distributions on the images directly and use those when possible reducing the most common setup time to ~0s. The 20 second improvement was just one example that I mentioned. With Python there are 9 cached versions while with Java we have 2 that we aren't even using now. The more we add the benefits will increase significantly. We can't cache everything because there is a limited amount of storage on our runners and there is a host of other software that users want cached. So fundamentally the problem we're trying to solve is **optimizing for the most common use cases across the board**

And there's again the profound confusion between `preinstalled` and `cache`. I don't know what you're basing those 20s on, but with #84 that would come down to probably < 1s. At the benefit of doing what you expect and having smaller, easier maintainable VM images.

> > It should not have an "and I really mean it" flag that defaults to "off".
> 
> The most common asks that we've overwhelmingly seen are people wanting faster builds which is why the default is off. If you're an open source repo or organization that really needs the latest updates of Java as soon as possible **always**, then you an easily opt-in, but most users that we've seen don't fall into that bucket. There will be documentation added for this like in `setup-python` that should clear any confusion about how we determine what is used.
> 
> > Since this is definitely a breaking change being proposed in v2 preview, and a dramatic modification to the common use cases and how OpenJDK updates will flow through to testing
> 
> I would like to reiterate that characterizing this change as dramatic is a bit of an overstatement. It's a new release, so it's acceptable to make any breaking changes. If you follow our migration guide and you're currently using `v1` with `zulu`, this change has no impact on hosted runners because there are no `zulu` distributions that are cached. So `checks-latest` being either `true` or `false` makes no difference whatsoever since a fresh VM will always require a download to take place. Our other actions like `setup-node` also have a checks-latest flag that defaults to `false` so this is a pattern we are implementing across the board as it best help optimize the action for the majority of users: https://github.com/actions/setup-node/blob/5c355be17065acf11598c7a9bb47112fbcf2bbdc/action.yml#L12-L14